@@ -2,6 +2,7 @@
   "name": "functions",
   "description": "Cloud Functions for Firebase",
   "scripts": {
+    "build": "tslint --project tsconfig.json ",
     "lint": "eslint .",
     "serve": "firebase serve --only functions",
     "shell": "firebase functions:shell",
@@ -10,12 +11,14 @@
     "logs": "firebase functions:log"
   },
   "engines": {
-    "node": "8"
+    "node": "10"
   },
   "dependencies": {
     "firebase": "^7.8.1",
     "firebase-admin": "^8.6.0",
     "firebase-functions": "^3.3.0",
+    "tslint": "^5.20.0",
+    "typescript": "^3.8.2",
     "uuid-token-generator": "^1.0.0"
   },
   "devDependencies": {
